requirements.txt
appdirs==1.4.4
appnope==0.1.2
autowrapt==1.0
backcall==0.2.0
boto3==1.17.53
botocore==1.20.53
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==5.0.7
distlib==0.3.1
env==0.1.0
epsagon==1.65.0
filelock==3.0.12
Flask==1.1.2
future==0.18.2
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.3.7
six==1.15.0
traitlets==5.0.5
urllib3==1.26.4
virtualenv==20.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
